map will change

+++ android/src/com/matji/sandwich/http/util/FileCache.java
@@ -37,8 +37,11 @@ public class FileCache {
     
     public void clear(){
         File[] files=cacheDir.listFiles();
-        for(File f:files)
-            f.delete();
+        if (files != null)
+        {
+            for(File f:files)
+                f.delete();
+        }
     }
 
 }
